Angie dev

+++ contracts/Test-helpers/ITestERC20.sol
@@ -4,7 +4,7 @@ pragma solidity ^0.8.9;
 import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 
 interface ITestERC20 is IERC20 {
-    function mint(address account, uint amount) external;
+    function mint(address account, uint256 amount) external;
 
-    function burn(address account, uint amount) external;
+    function burn(address account, uint256 amount) external;
 }
