+++ contracts/interfaces/IAmericanBinaryOptions.sol
@@ -2,46 +2,49 @@
 pragma solidity 0.8.9;
 
 interface IAmericanBinaryOptions {
-
     struct PricesDeltaStdVega {
-        uint callPrice;
-        uint putPrice;
+        uint256 callPrice;
+        uint256 putPrice;
         int callDelta;
         int putDelta;
-        uint stdVega;
+        uint256 stdVega;
     }
 
-    function mockCalculation(uint coverage, uint duration, uint risk) external pure returns(uint premium);
+    function mockCalculation(
+        uint256 coverage,
+        uint256 duration,
+        uint256 risk
+    ) external pure returns (uint256 premium);
 
     function americanBinaryOptionPrices(
-        uint timeToExpirySec,
-        uint volatilityDecimal,
-        uint spotDecimal,
-        uint strikeDecimal,
+        uint256 timeToExpirySec,
+        uint256 volatilityDecimal,
+        uint256 spotDecimal,
+        uint256 strikeDecimal,
         int rateDecimal
     ) external view returns (int);
 
-    function abs(int x) external pure returns (uint);
+    function abs(int x) external pure returns (uint256);
 
-    function exp(uint x) external pure returns (uint);
+    function exp(uint256 x) external pure returns (uint256);
 
-    function exp(int x) external pure returns (uint);
+    function exp(int x) external pure returns (uint256);
 
-    function sqrt(uint x) external pure returns (uint y);
+    function sqrt(uint256 x) external pure returns (uint256 y);
 
     function optionPrices(
-        uint timeToExpirySec,
-        uint volatilityDecimal,
-        uint spotDecimal,
-        uint strikeDecimal,
+        uint256 timeToExpirySec,
+        uint256 volatilityDecimal,
+        uint256 spotDecimal,
+        uint256 strikeDecimal,
         int rateDecimal
-    ) external pure returns (uint call, uint put);
+    ) external pure returns (uint256 call, uint256 put);
 
     function pricesDeltaStdVega(
-        uint timeToExpirySec,
-        uint volatilityDecimal,
-        uint spotDecimal,
-        uint strikeDecimal,
+        uint256 timeToExpirySec,
+        uint256 volatilityDecimal,
+        uint256 spotDecimal,
+        uint256 strikeDecimal,
         int rateDecimal
     ) external view returns (PricesDeltaStdVega memory);
 }
